+++ src/interfaces/store/projectStore.interface.ts
@@ -2,7 +2,7 @@ import { Project } from "@type/models";
 import { ServiceResponse } from "@type/services.types";
 
 export interface ProjectStore {
-	createProject: (isDefault?: boolean) => ServiceResponse<{ name: string; projectId: string }>;
+	createProject: (name: string, isDefault?: boolean) => ServiceResponse<{ name: string; projectId: string }>;
 	deleteProject: (projectId: string) => ServiceResponse<undefined>;
 	getProject: (projectId: string) => ServiceResponse<Project>;
 	getProjectsList: () => ServiceResponse<Project[]>;
